Get description & images from STEX api instead of scraping
